backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS
python-dotenv==1.0.0

# AI/ML (나중에 사용)
# openai==1.10.0
# langchain==0.1.0

# Utils
python-dateutil==2.8.2
